tpm2: Cast '1' to UINT64 before shifting (never shifts more than 13 bytes)
Cast '1' to UINT64 before potentially shifting by 64 bit. However, none of
the SHA1 and SHA2-family algorithm identifiers cause a 64bit shift. Their
values are 4 (SHA1), 11 (SHA256), 12 (SHA384) and 13 (SHA512). SM3_256 is
not enabled and would have value 18.

+++ src/tpm2/NVMarshal.c
@@ -931,7 +931,7 @@ PCR_SAVE_Unmarshal(PCR_SAVE *data, BYTE **buffer, INT32 *size,
         }
         if (t) {
             if (rc == TPM_RC_SUCCESS) {
-                algs_needed &= ~(1 << algid);
+                algs_needed &= ~((UINT64)1 << algid);
                 rc = UINT16_Unmarshal(&array_size, buffer, size);
             }
             if (rc == TPM_RC_SUCCESS && array_size != needed_size) {
@@ -1113,7 +1113,7 @@ PCR_Unmarshal(PCR *data, BYTE **buffer, INT32 *size,
         }
         if (t) {
             if (rc == TPM_RC_SUCCESS) {
-                algs_needed &= ~(1 << algid);
+                algs_needed &= ~((UINT64)1 << algid);
                 rc = UINT16_Unmarshal(&array_size, buffer, size);
             }
             if (rc == TPM_RC_SUCCESS && array_size != needed_size) {
